Adds header, footer
Added the bulk of the page elements, and styled them as best I could. Many hacks involved, most notably the use of CSS grid to center the logo and searchbar. Definitely do not know what I'm doing when it comes to layout. Tried pretty hard to match the google account avatar in the top right, but definitely a challenge. Didn't even try to add the google apps thing.